extract ProductCard
